refactor: rename linestring to coordinate
