+++ src/infrastructure/external/srtm.rs
@@ -8,7 +8,7 @@ use std::path::{Path, PathBuf};
 use byteorder::{BigEndian, ByteOrder, LittleEndian, ReadBytesExt};
 use num_traits::FromPrimitive;
 
-use crate::domain::model::linestring::Coordinate;
+use crate::domain::model::coordinate::Coordinate;
 use crate::domain::model::types::{Elevation, Latitude, Longitude};
 use crate::domain::repository::ElevationApi;
 use crate::utils::error::{ApplicationError, ApplicationResult};
